refactor(cars): 우승자를 판단하는 기능의 책임을 Referee에서 Cars로 위임
