+++ docs/README.md
@@ -12,9 +12,6 @@
 
 - [x] 경주할 자동차의 정보를 가진다.
 - [x] 이름으로 자동차를 조회할 수 있다.
-
-### Referee
-
 - [ ] 우승자를 판단할 수 있다.
 
 ## View
