--- src/main/java/domain/Referee.java
@@ -1,7 +1,0 @@
-package domain;
-
-public class Referee {
-    public static Cars judgeWinner(Cars cars) {
-        return null;
-    }
-}
